feat(cli): simplified the arguments after some testing
After a few iterations, and comparison to other commands I realised I was over complicating the command arguments, and replacing features already in bash, being path expansion. So with this in mind I update the arguments to be more idiomatic of a bash command.

Only took a few tries to get right... :facepalm: